HBX-1648: Move class 'org.hibernate.tool.hbm2x.HibernateMappingExporter' to 'org.hibernate.tool.internal.export.hbm.HibernateMappingExporter'
